+++ extension/background/main.js
@@ -251,9 +251,29 @@ const openWakeword = util.serializeCalls(async function() {
     await sendMessage({ type: "updateWakeword" }, { tabId: tab.id });
     await browserUtil.makeTabActive(activeTab.id);
   } else {
-    await sendMessage({ type: "updateWakeword" }, { tabId: tabs[0].id });
+    try {
+      await sendMessage({ type: "updateWakeword" }, { tabId: tabs[0].id });
+    } catch (e) {
+      if (e.message.includes("Could not establish connection")) {
+        await browser.tabs.reload(tabs[0].id);
+      }
+      throw e;
+    }
   }
   wakewordMaybeOpen = true;
+});
+
+registerHandler("focusWakewordTab", async (message, sender) => {
+  const tab = await browserUtil.activeTab();
+  // For some reason this takes a long time to return, too long, and
+  // doesn't return in time, so we let it run in the background and return
+  // the tab.id fast:
+  browserUtil.makeTabActive(sender.tab.id);
+  return tab.id;
+});
+
+registerHandler("unfocusWakewordTab", async message => {
+  await browserUtil.makeTabActive(message.tabId);
 });
 
 // These message handlers are kept in main.js to avoid cases where the module
@@ -327,7 +347,11 @@ settings.watch("enableWakeword", openWakeword);
 settings.watch("wakewords", openWakeword);
 settings.watch("wakewordSensitivity", openWakeword);
 
-openWakeword();
+// If the wakeword is opened too soon, there can be a permission-related race condition
+// where the wakeword page won't load:
+setTimeout(() => {
+  openWakeword();
+}, 500);
 
 function setUninstallURL() {
   const url = telemetry.createSurveyUrl(UNINSTALL_SURVEY);