+++ extension/intents/routines/routineExecutor.js
@@ -4,19 +4,22 @@ import * as intentRunner from "../../background/intentRunner.js";
 import { sendMessage } from "../../communicate.js";
 
 export let pausedRoutineExecutor = null;
+export let currentRoutineExecutor = null;
 export class RoutineExecutor {
   constructor(
     routineName,
     subcommands,
     programCounter = 0,
     states = null,
-    startLoopProgramCounter = 0
+    startLoopProgramCounter = 0,
+    stopRoutine = false
   ) {
     this.routineName = routineName;
     this.subcommands = subcommands;
     this.programCounter = programCounter;
     this.states = states;
     this.startLoopProgramCounter = startLoopProgramCounter;
+    this.stopRoutine = stopRoutine;
   }
 
   mapState(subcommand, state) {
@@ -86,14 +89,19 @@ export class RoutineExecutor {
       exc.displayMessage = "Another routine is already active.";
       throw exc;
     }
+    currentRoutineExecutor = this;
     for (
       ;
       this.programCounter < this.subcommands.length;
       this.programCounter++
     ) {
+      if (this.stopRoutine === true) {
+        return true;
+      }
       const subcommand = this.subcommands[this.programCounter];
       const stopRoutine = await this.runSubcommand(subcommand);
       if (stopRoutine === true) {
+        currentRoutineExecutor = null;
         return true;
       }
     }
@@ -102,6 +110,7 @@ export class RoutineExecutor {
       exc.displayMessage = "'End for' is required at the end of loop.";
       throw exc;
     }
+    currentRoutineExecutor = null;
     return true;
   }
 
@@ -141,5 +150,11 @@ export class RoutineExecutor {
       }, 1000);
       await this.continue();
     }, ms);
+  }
+
+  stop() {
+    this.stopRoutine = true;
+    pausedRoutineExecutor = null;
+    currentRoutineExecutor = null;
   }
 }