+++ extension/popup/voice.js
@@ -8,7 +8,9 @@ import { sendMessage } from "../communicate.js";
 const STT_SERVER_URL =
   buildSettings.sttServer || "https://speaktome-2.services.mozilla.com";
 const DEEP_SPEECH_URL = buildSettings.deepSpeechServer;
-const LANGUAGE = "en-US";
+
+const locale = settings.getSettings().userLocale || navigator.language;
+const LANGUAGE = locale && locale.startsWith("en-") ? locale : "en-US";
 
 export class Recorder {
   constructor(stream) {